requirements.txt
langgraph==0.2.28
langchain==0.3.1
langchain-core==0.3.6
langchain-google-genai==2.0.0
langchain-community==0.3.1
pypdf==4.3.1
python-dotenv==1.0.1
pandas==2.2.0
tabulate==0.9.0